api/requirements.txt
python-telegram-bot
httpx
fastapi
uvicorn